Refactored sub-package demo to util in test/../io/importer/labelimage
